requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.123.4
ffmpeg-python==0.2.0
filelock==3.20.0
Flask==2.3.3
flask-cors==6.0.1
fsspec==2025.10.0
future==1.0.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.45.1
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.4.2
numba==0.62.1
numpy==2.2.6
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
opencc-python-reimplemented==0.1.7
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
python-multipart==0.0.20
regex==2025.11.3
requests==2.32.5
semantic-version==2.10.0
setuptools-rust==1.12.0
starlette==0.50.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.1
torchaudio==2.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.4
